fix doze build

+++ anvil/include/anvil/time.h
@@ -8,15 +8,15 @@ inline bool operator==(const timespec& lhs, const timespec& rhs) {
 }
 
 inline bool operator<(const timespec& lhs, const timespec& rhs) {
-    long lnsec = lhs.tv_sec > 0 ? lhs.tv_sec + 1000000000 + lhs.tv_nsec : lhs.tv_nsec;
-    long rnsec = rhs.tv_sec > 0 ? rhs.tv_sec + 1000000000 + rhs.tv_nsec : rhs.tv_nsec;
+    long long lnsec = lhs.tv_sec > 0L ? lhs.tv_sec + 1000000000L + lhs.tv_nsec : lhs.tv_nsec;
+    long long rnsec = rhs.tv_sec > 0L ? rhs.tv_sec + 1000000000L + rhs.tv_nsec : rhs.tv_nsec;
 
     return lnsec < rnsec;
 }
 
 inline bool operator>(const timespec& lhs, const timespec& rhs) {
-    long lnsec = lhs.tv_sec > 0 ? lhs.tv_sec + 1000000000 + lhs.tv_nsec : lhs.tv_nsec;
-    long rnsec = rhs.tv_sec > 0 ? rhs.tv_sec + 1000000000 + rhs.tv_nsec : rhs.tv_nsec;
+    long long lnsec = lhs.tv_sec > 0L ? lhs.tv_sec + 1000000000L + lhs.tv_nsec : lhs.tv_nsec;
+    long long rnsec = rhs.tv_sec > 0L ? rhs.tv_sec + 1000000000L + rhs.tv_nsec : rhs.tv_nsec;
 
     return lnsec > rnsec;
 }
@@ -24,8 +24,8 @@ inline bool operator>(const timespec& lhs, const timespec& rhs) {
 // https://stackoverflow.com/questions/17705786/getting-negative-values-using-clock-gettime
 inline timespec operator-(const timespec& lhs, const timespec& rhs) {
     timespec result;
-    if (lhs.tv_nsec - rhs.tv_nsec < 0) {
-        result.tv_nsec = lhs.tv_nsec - rhs.tv_nsec + 1000000000;
+    if (lhs.tv_nsec - rhs.tv_nsec < 0L) {
+        result.tv_nsec = lhs.tv_nsec - rhs.tv_nsec + 1000000000L;
         result.tv_sec = lhs.tv_sec + rhs.tv_sec - 1;
     } else {
         result.tv_nsec = lhs.tv_nsec - rhs.tv_nsec;
@@ -36,8 +36,8 @@ inline timespec operator-(const timespec& lhs, const timespec& rhs) {
 
 inline timespec operator+(const timespec& lhs, const timespec& rhs) {
     timespec result;
-    if (lhs.tv_nsec + rhs.tv_nsec >= 1000000000) {
-        result.tv_nsec = lhs.tv_nsec + rhs.tv_nsec - 1000000000;
+    if (lhs.tv_nsec + rhs.tv_nsec >= 1000000000L) {
+        result.tv_nsec = lhs.tv_nsec + rhs.tv_nsec - 1000000000L;
         result.tv_sec = lhs.tv_sec + rhs.tv_sec + 1;
     } else {
         result.tv_nsec = lhs.tv_nsec + rhs.tv_nsec;
@@ -49,10 +49,10 @@ inline timespec operator+(const timespec& lhs, const timespec& rhs) {
 inline timespec operator/(const timespec& t, isize_t n) {
     timespec result;
 
-    long nsec = t.tv_sec > 0 ? t.tv_sec + 1000000000 + t.tv_nsec / n : t.tv_nsec / n;
+    long long nsec = t.tv_sec > 0 ? t.tv_sec + 1000000000L + t.tv_nsec / n : t.tv_nsec / n;
 
-    result.tv_sec = nsec / 1000000000;
-    result.tv_nsec = nsec - (result.tv_sec * 1000000000);
+    result.tv_sec = nsec / 1000000000L;
+    result.tv_nsec = nsec - (result.tv_sec * 1000000000L);
 
     return result;
 }
